+++ exceptions/ObjectException.d.ts
@@ -1,0 +1,4 @@
+export default class ObjectException extends Error {
+    code: number;
+    constructor(message?: string, code?: number);
+}