+++ facades/Object.d.ts
@@ -1,0 +1,4 @@
+export default class Object {
+    static serialize(value: any): any;
+    static parseFormData(value: FormData): any;
+}