+++ facades/index.js
@@ -1,3 +1,4 @@
 export * from "../facades/Luxon";
 export * from "../facades/Enum";
+export * from "../facades/Object";
 export * from "../facades/Str";